+++ xpdacq/tests/test_devices.py
@@ -1,0 +1,6 @@
+import xpdacq.devices as devices
+
+
+def test_CalibrationData():
+    """Test the class can be constructed"""
+    devices.CalibrationData(name="calib")