+++ news/news.rst
@@ -1,8 +1,12 @@
 **Added:**
 
-* Add `xpdacq.plan_stubs.mv_and_trigger_and_read` for the multi-detector plan for both XRD and PDF data collection in one run.
+* Add `xpdacq.factory.BasicPlans` for the multi-detector plan for both XRD and PDF data collection in one run.
+
+* Add `xpdacq.factory.MultiDistPlans` for the moving detector measurement for both XRD and PDF data collection in one run.
 
 * Add `xpdacq.devices.CalibrationData`, a class to store the calibration data of a detector in configuration attributes.
+
+
 
 **Changed:**